# swaggerDocument.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "SPCA API"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/centres": {
      "get": {
        "tags": ["Centre"],
        "summary": "Retrieves a list of all centres",
        "responses": {
          "200": {
            "description": "A list of centres",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Centre"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Centre"],
        "summary": "Creates a new centre",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewCentre"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Centre created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Centre"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/centres/{centreId}": {
      "parameters": [
        {
          "name": "centreId",
          "in": "path",
          "description": "ID of the centre to retrieve",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "tags": ["Centre"],
        "summary": "Retrieves a single centre by ID",
        "responses": {
          "200": {
            "description": "Centre found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Centre"
                }
              }
            }
          },
          "404": {
            "description": "Centre not found"
          }
        }
      }
    },
    "/centres/{centreId}/dogs": {
      "parameters": [
        {
          "name": "centreId",
          "in": "path",
          "description": "ID of the centre to retrieve dogs from",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "tags": ["Dog"],
        "summary": "Retrieves a list of all dogs in a specific centre",
        "responses": {
          "200": {
            "description": "A list of dogs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Dog"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Centre not found"
          }
        }
      },
      "post": {
        "tags": ["Dog"],
        "summary": "Adds a new dog to a specific centre",
        "parameters": [
          {
            "name": "centreId",
            "in": "path",
            "description": "ID of the centre to add the dog to",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Dog object to be added to the centre",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewDog"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dog added to the centre",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "uri",
                  "example": "https://localhost:8080/centres/1/dogs/1"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input, object invalid"
          },
          "404": {
            "description": "Centre not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/centres/{centreId}/dogs/{dogId}": {
      "parameters": [
        {
          "name": "centreId",
          "in": "path",
          "description": "ID of the centre the dog belongs to",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "dogId",
          "in": "path",
          "description": "ID of the dog to retrieve",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "tags": ["Dog"],
        "summary": "Retrieves a single dog by ID within a specific centre",
        "responses": {
          "200": {
            "description": "Dog found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dog"
                }
              }
            }
          },
          "404": {
            "description": "Dog / centre not found"
          }
        }
      },
      "delete": {
        "tags": ["Dog"],
        "summary": "Removes a dog from a specific centre",
        "parameters": [
          {
            "name": "centreId",
            "in": "path",
            "description": "ID of the centre the dog belongs to",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dogId",
            "in": "path",
            "description": "ID of the dog to remove",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Dog removed successfully"
          },
          "404": {
            "description": "Centre or dog not found"
          }
        }
      }
    },
...
    "/centres/{centreId}/dogs/{dogId}/weight": {
      "post": {
        "summary": "Update the weight of a dog",
        "tags": ["Dog"],
        "parameters": [
          {
            "name": "centreId",
            "in": "path",
            "description": "ID of the centre the dog belongs to",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dogId",
            "in": "path",
            "description": "ID of the dog to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "The weight of the dog on the scale",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewWeight"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Dog weight updated successfully"
          },
          "400": {
            "description": "Invalid input, object invalid"
          },
          "404": {
            "description": "Centre or dog not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Centre": {
        "type": "object",
        "required": ["id", "name", "location"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the centre"
          },
          "name": {
            "type": "string",
            "description": "Name of the centre"
          },
          "location": {
            "type": "string",
            "description": "Location of the centre"
          },
          "dogs": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "ID of a dog at the centre"
            }
          },
          "scales": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "ID of a scale at the centre"
            }
          }
        }
      },
      "NewCentre": {
        "type": "object",
        "required": ["name", "location"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the centre"
          },
          "location": {
            "type": "string",
            "description": "Location of the centre"
          }
        }
      },
...
      "Dog": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DogDTO"
          }
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the dog"
          },
          "weight": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Weight"
            }
          },
          "centreId": {
            "type": "string",
            "description": "ID of the centre the dog is associated with"
          }
        }
      },
      "NewDog": {
        "type": "object",
        "required": ["name", "description"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the dog"
          },
          "description": {
            "type": "string",
            "description": "Description of the dog"
          },
          "image": {
            "type": "string",
            "description": "URL of the dog's image"
          }
        }
      },
      "DogDTO": {
        "allOf": [
          {
            "$ref": "#/components/schemas/NewDog"
          }
        ],
        "properties": {
          "idBeingWeighted": {
            "type": "boolean",
            "description": "Whether the dog is currently being weighed"
          }
        }
      },
      "NewWeight": {
        "type": "object",
        "properties": {
          "weight": {
            "type": "number",
            "description": "The weight of the dog on the scale",
            "format": "float",
            "minimum": 0.1
          }
        },
        "required": ["weight"]
      },
      "Weight": {
        "allOf": [
          {
            "$ref": "#/components/schemas/NewWeight"
          }
        ],
        "properties": {
          "dateWeighed": {
            "type": "string",
            "description": "The date and time that the dog was weighed",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}